+++ AB460Proniya/Areas/admin/Controllers/ColorController.cs
@@ -102,7 +102,7 @@ namespace AB460Proniya.Areas.admin.Controllers
             return RedirectToAction(nameof(Index));
         }
         [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(int id, bool confirim)
         {
             if (id <= 0) return BadRequest();
 
@@ -111,10 +111,19 @@ namespace AB460Proniya.Areas.admin.Controllers
 
             if (existed is null) return NotFound();
 
-            _context.Colors.Remove(existed);
+            if (confirim)
+            {
+                _context.Colors.Remove(existed);
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+
+                return View(existed);
+            }
+
         }
         [Authorize(Roles = "Admin,Moderator")]
         public async Task<IActionResult> Details(int id)
